# docs/openapi.yaml
openapi: 3.2.0
info:
  title: Bill Payment API
  description: |
    A comprehensive RESTful API for bill payment operations, designed for API automation practice.

    ## Overview
    This API provides endpoints to manage billers, bills, payments, payment methods, users, and file uploads.
    It supports various authentication methods, content types, and includes realistic business logic.

    ## Authentication
    The API supports multiple authentication methods. All endpoints except `/health` require authentication.

    ### API Key (Recommended for Testing)
    - **Header**: `X-API-Key: demo-api-key-123`
    - **Query Parameter**: `?api_key=demo-api-key-123`

    ### Bearer Token (JWT)
    ```
    Authorization: Bearer demo-jwt-token-456
    ```

    ### Basic Authentication
    ```
    Authorization: Basic ZGVtbzpwYXNzd29yZDEyMw==
    ```
    (Credentials: `demo:password123`)

    ### Cookie-Based Session (NEW)
    ```
    Cookie: session_id=demo-session-abc123
    ```
    Session cookies are validated for authentication. Use the demo session token above for testing.

    ### OAuth2 Client Credentials
    ```bash
    POST /oauth/token
    Content-Type: application/x-www-form-urlencoded

    grant_type=client_credentials&client_id=demo-client&client_secret=demo-secret-789
    ```

    ## Content Types
    The API supports multiple content types:

    ### JSON (Default)
    - Request: `Content-Type: application/json`
    - Response: `Accept: application/json`

...

...

...

...

...

    ```
...
servers:
...
paths:
...
  /v1/billers:
    get:
      tags: [Billers]
      summary: List all billers
      description: |
        Get a paginated list of all billers (service providers).
        Supports filtering by category, active status, and search.
      operationId: listBillers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: category
          in: query
          description: Filter by biller category
          schema:
            $ref: '#/components/schemas/BillerCategory'
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: search
          in: query
          description: Search in name, display name, description
          schema:
            type: string
            example: airtel
        - name: fetch_bill_supported
          in: query
          description: Filter billers that support bill fetching
          schema:
            type: boolean
      responses:
        '200':
          description: List of billers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillersListResponse'
              examples:
                telecomBillers:
                  summary: Telecom billers
                  value:
                    success: true
                    data:
                      - id: biller-airtel-postpaid
                        name: airtel
                        displayName: Airtel Postpaid
                        category: telecom
                        minAmount: 1
                        maxAmount: 50000
                        isActive: true
                    meta:
                      requestId: "abc-123"
                      timestamp: "2025-12-18T10:30:00.000Z"
                      version: v1
                      pagination:
                        page: 1
                        limit: 10
                        total: 25
                        totalPages: 3
                        hasNext: true
                        hasPrev: false
                        offset: 0
                        nextCursor: null
                        prevCursor: null
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Billers]
      summary: Create a new biller
      description: Create a new biller/service provider. Typically for admin use.
      operationId: createBiller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillerInput'
            examples:
              minimalBiller:
                summary: Minimal required fields
                value:
                  name: newbiller
                  displayName: New Biller Inc.
                  category: telecom
              fullBiller:
                summary: All fields
                value:
                  name: newbiller
                  displayName: New Biller Inc.
                  category: electricity
                  description: A new electricity provider
                  customerIdLabel: Account Number
                  customerIdPattern: "^[0-9]{10}$"
                  minAmount: 100
                  maxAmount: 50000
                  fetchBillSupported: true
                  partialPaymentAllowed: true
                  supportedPaymentModes: [upi, credit_card, debit_card]
      responses:
        '201':
          description: Biller created
          headers:
            Location:
              schema:
                type: string
              description: URL of the created biller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillerResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
    additionalOperations:
      # OpenAPI 3.2 describes non-standard verbs here, keyed by the uppercase
      # method name. This is the native RFC 10008 form; the POST fallback at
      # /v1/billers/query documents the same operation for tooling that does
      # not yet render additionalOperations.
      QUERY:
        tags: [HTTP QUERY]
        summary: Search billers with a request body (native QUERY)
        description: |
          Native `QUERY /v1/billers` - RFC 10008. Safe, idempotent, and
          carries a JSON body. Identical in behaviour to
          `POST /v1/billers/query`; `meta.querySource` distinguishes them.
        operationId: queryBillersNative
        requestBody:
          required: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRequest'
              example:
                filter: { category: telecom, isActive: true }
                sort: [displayName]
        responses:
          '200':
            description: Matching resources
            headers:
              Accept-Query:
                schema: { type: string }
              ETag:
                schema: { type: string }
              Link:
                schema: { type: string }
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/QueryResponse'
          '304':
            description: Not modified
          '400':
            description: Content-Type missing, or body inconsistent with it
          '415':
            description: Unsupported Content-Type
          '422':
            description: Well-formed body, but the query cannot be processed
...
  /v1/bills:
    get:
      tags: [Bills]
      summary: List bills
      description: Get a paginated list of bills with optional filters
      operationId: listBills
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
            example: user-demo-001
        - name: biller_id
          in: query
          description: Filter by biller ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by bill status
          schema:
            $ref: '#/components/schemas/BillStatus'
        - name: created_by
          in: query
          description: Filter by creator (API key identifier)
          schema:
            type: string
        - name: due_before
          in: query
          description: Filter bills due before this date
          schema:
            type: string
            format: date
        - name: due_after
          in: query
          description: Filter bills due after this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of bills
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Bills]
      summary: Create a new bill
      description: Register a new bill for payment
      operationId: createBill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillInput'
            examples:
              simpleBill:
                summary: Simple bill
                value:
                  userId: user-demo-001
                  billerId: biller-airtel-postpaid
                  customerIdentifier: "9876543210"
                  amount: 599.00
              fullBill:
                summary: Bill with all details
                value:
                  userId: user-demo-001
                  billerId: biller-tata-power
                  customerIdentifier: "1234567890"
                  customerName: John Doe
                  nickname: Home Electricity
                  amount: 2500.00
                  dueDate: "2025-12-25"
                  billDate: "2025-12-01"
                  billPeriod: "Dec 2025"
                  status: pending
      responses:
        '201':
          description: Bill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: User or biller not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    additionalOperations:
      # OpenAPI 3.2 describes non-standard verbs here, keyed by the uppercase
      # method name. This is the native RFC 10008 form; the POST fallback at
      # /v1/bills/query documents the same operation for tooling that does
      # not yet render additionalOperations.
      QUERY:
        tags: [HTTP QUERY]
        summary: Search bills with a request body (native QUERY)
        description: |
          Native `QUERY /v1/bills` - RFC 10008. Safe, idempotent, and
          carries a JSON body. Identical in behaviour to
          `POST /v1/bills/query`; `meta.querySource` distinguishes them.
        operationId: queryBillsNative
        requestBody:
          required: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRequest'
              example:
                filter: { status: { in: [pending, overdue] }, amount: { gte: 500 } }
                sort: ["-amount"]
                limit: 10
        responses:
          '200':
            description: Matching resources
            headers:
              Accept-Query:
                schema: { type: string }
              ETag:
                schema: { type: string }
              Link:
                schema: { type: string }
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/QueryResponse'
          '304':
            description: Not modified
          '400':
            description: Content-Type missing, or body inconsistent with it
          '415':
            description: Unsupported Content-Type
          '422':
            description: Well-formed body, but the query cannot be processed
...
  /v1/payment-methods:
    get:
      tags: [Payment Methods]
      summary: List payment methods
      operationId: listPaymentMethods
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: user_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/PaymentMethodType'
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of payment methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodsListResponse'

    post:
      tags: [Payment Methods]
      summary: Create payment method
      operationId: createPaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodInput'
            examples:
              upi:
                summary: UPI Payment Method
                value:
                  userId: user-demo-001
                  type: upi
                  displayName: My UPI
                  upiId: myupi@okaxis
                  isDefault: true
              creditCard:
                summary: Credit Card
                value:
                  userId: user-demo-001
                  type: credit_card
                  displayName: HDFC Credit Card
                  cardLastFour: "1234"
                  cardNetwork: visa
                  cardExpiryMonth: 12
                  cardExpiryYear: 2027
                  cardHolderName: JOHN DOE
      responses:
        '201':
          description: Payment method created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponse'
...
  /v1/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: kyc_status
          in: query
          schema:
            $ref: '#/components/schemas/KycStatus'
        - name: search
          in: query
          description: Search by email, name, or phone
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'

    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
            examples:
              simpleUser:
                summary: Minimal user
                value:
                  email: newuser@example.com
                  firstName: New
                  lastName: User
              fullUser:
                summary: User with all details
                value:
                  email: newuser@example.com
                  phone: "+919876543210"
                  firstName: New
                  lastName: User
                  kycStatus: pending
                  address:
                    line1: 123 Main Street
                    city: Mumbai
                    state: Maharashtra
                    postalCode: "400001"
                    country: IN
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: User with this email already exists
    additionalOperations:
      # OpenAPI 3.2 describes non-standard verbs here, keyed by the uppercase
      # method name. This is the native RFC 10008 form; the POST fallback at
      # /v1/users/query documents the same operation for tooling that does
      # not yet render additionalOperations.
      QUERY:
        tags: [HTTP QUERY]
        summary: Search users with a request body (native QUERY)
        description: |
          Native `QUERY /v1/users` - RFC 10008. Safe, idempotent, and
          carries a JSON body. Identical in behaviour to
          `POST /v1/users/query`; `meta.querySource` distinguishes them.
        operationId: queryUsersNative
        requestBody:
          required: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRequest'
              example:
                filter: { kycStatus: verified }
                fields: [id, email, kycStatus]
        responses:
          '200':
            description: Matching resources
            headers:
              Accept-Query:
                schema: { type: string }
              ETag:
                schema: { type: string }
              Link:
                schema: { type: string }
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/QueryResponse'
          '304':
            description: Not modified
          '400':
            description: Content-Type missing, or body inconsistent with it
          '415':
            description: Unsupported Content-Type
          '422':
            description: Well-formed body, but the query cannot be processed
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed in header
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: API key passed as query parameter
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication (username:password)
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: |
        Session cookie authentication. 
        Demo value: `session_id=demo-session-abc123`
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            read:all: Read access to all resources
            write:all: Write access to all resources
        password:
          tokenUrl: /oauth/token
          scopes:
            read:all: Read access
            write:all: Write access
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: MISSING_AUTHENTICATION
              message: Authentication required. Provide API Key, Bearer token, or Basic auth credentials.
              traceId: abc-123
              timestamp: "2025-12-18T10:30:00.000Z"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid input data
              traceId: abc-123
              timestamp: "2025-12-18T10:30:00.000Z"
              details:
                - field: email
                  code: REQUIRED
                  message: email is required
    Conflict:
      description: Conflict - resource already exists or cannot be deleted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    QueryRequest:
      type: object
      description: |
        The body of an HTTP QUERY request. Every member is optional - an empty
        object matches everything.
      properties:
        filter:
          type: object
          description: |
            Field conditions, combined with AND unless nested under `or`/`not`.
            A bare value is shorthand for `eq`, so `{"status": "pending"}` and
            `{"status": {"eq": "pending"}}` are the same. Nesting is limited to
            four levels.
          additionalProperties: true
          example:
            status: { in: [pending, overdue] }
            amount: { gte: 500, lte: 10000 }
            or:
              - billerCategory: telecom
              - billerCategory: electricity
        sort:
          type: array
          description: Field names, prefixed with `-` for descending.
          items: { type: string }
          example: ["-amount", "dueDate"]
        fields:
          type: array
          description: Sparse fieldset. `id` is always included.
          items: { type: string }
          example: [id, amount, status]
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        offset:
          type: integer
          minimum: 0
          description: Alternative to `page`.
    QueryResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items: { type: object }
        meta:
          type: object
          properties:
            requestId: { type: string }
            timestamp: { type: string, format: date-time }
            version: { type: string }
            pagination:
              $ref: '#/components/schemas/Pagination'
            query:
              type: object
              description: The query as the server understood it, echoed back.
            querySource:
              type: string
              enum: [query-method, method-override, query-path]
...
    Address:
      type: object
      properties:
        line1:
          type: string
          example: 123 Main Street
        line2:
          type: string
        city:
          type: string
          example: Mumbai
        state:
          type: string
          example: Maharashtra
        postalCode:
          type: string
          example: "400001"
        country:
          type: string
          default: IN
          example: IN
...
    ResponseMeta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          example: v1
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
            traceId:
              type: string
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
...
    BillerCategory:
      type: string
      enum:
        - telecom
        - electricity
        - water
        - gas
        - broadband
        - dth
        - insurance
        - credit_card
        - loan
        - municipal_tax
    BillStatus:
      type: string
      enum: [pending, paid, overdue, partially_paid, cancelled]
...
    PaymentMethodType:
      type: string
      enum: [upi, credit_card, debit_card, net_banking, wallet]
    KycStatus:
      type: string
      enum: [pending, verified, rejected]

    # Biller schemas
...
    BillerInput:
      type: object
      required:
        - name
        - displayName
        - category
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        displayName:
          type: string
          minLength: 2
          maxLength: 100
        category:
          $ref: '#/components/schemas/BillerCategory'
        logoUrl:
          type: string
          format: uri
        description:
          type: string
          maxLength: 500
        supportedPaymentModes:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethodType'
        fetchBillSupported:
          type: boolean
          default: true
        partialPaymentAllowed:
          type: boolean
          default: false
        minAmount:
          type: number
          minimum: 0
          default: 1
        maxAmount:
          type: number
          default: 100000
        customerIdLabel:
          type: string
          default: Customer ID
        customerIdPattern:
          type: string
        isActive:
          type: boolean
          default: true
...
    BillerResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Biller'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    BillersListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Biller'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    BillInput:
      type: object
      required:
        - userId
        - billerId
        - customerIdentifier
        - amount
      properties:
        userId:
          type: string
        billerId:
          type: string
        customerIdentifier:
          type: string
          minLength: 4
          maxLength: 20
        customerName:
          type: string
        nickname:
          type: string
          maxLength: 50
        amount:
          type: number
          minimum: 1
        currency:
          type: string
          default: INR
        dueDate:
          type: string
          format: date
        billDate:
          type: string
          format: date
        billPeriod:
          type: string
        status:
          $ref: '#/components/schemas/BillStatus'
...
    BillResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Bill'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    BillsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Bill'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    PaymentMethodInput:
      type: object
      required:
        - userId
        - type
        - displayName
      properties:
        userId:
          type: string
        type:
          $ref: '#/components/schemas/PaymentMethodType'
        displayName:
          type: string
          minLength: 2
          maxLength: 50
        isDefault:
          type: boolean
          default: false
        upiId:
          type: string
          description: Required for type=upi
        cardLastFour:
          type: string
          pattern: "^[0-9]{4}$"
          description: Required for card types
        cardNetwork:
          type: string
          enum: [visa, mastercard, amex, rupay, discover]
        cardExpiryMonth:
          type: integer
          minimum: 1
          maximum: 12
        cardExpiryYear:
          type: integer
          minimum: 2024
        cardHolderName:
          type: string
        bankName:
          type: string
        bankAccountLastFour:
          type: string
        walletProvider:
          type: string
        walletId:
          type: string
    PaymentMethodResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PaymentMethod'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PaymentMethodsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethod'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    UserInput:
      type: object
      required:
        - email
        - firstName
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: "^\\+?[0-9]{10,15}$"
        firstName:
          type: string
          minLength: 1
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        kycStatus:
          $ref: '#/components/schemas/KycStatus'
        address:
          $ref: '#/components/schemas/Address'
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
security:
  - ApiKeyHeader: []
  - ApiKeyQuery: []
  - BearerAuth: []
  - BasicAuth: []
  - CookieAuth: []
  - OAuth2: []